test(deps): move install.rs tests to dedicated install_tests.rs module
Refactor dependency-install status refresh tests by relocating them
from installer/src/deps/install.rs into a new installer/src/deps/install_tests.rs
file. This cleanup improves code organization and separation of test code
from implementation code without behavior changes.
