+++ installer/src/deps/install.rs
@@ -331,157 +331,5 @@ fn update_status_after_install(
 }
 
 #[cfg(test)]
-mod tests {
-    //! Tests for dependency-install status refresh behaviour.
-
-    use super::*;
-
-    #[test]
-    fn update_status_after_install_refreshes_link_for_local_cargo_dylint_installs() {
-        let missing_link = DylintToolStatus {
-            cargo_dylint: false,
-            dylint_link: false,
-        };
-        let executor =
-            crate::test_utils::StubExecutor::new(vec![crate::test_utils::ExpectedCall {
-                cmd: "dylint-link",
-                args: vec!["--version"],
-                result: Ok(crate::test_utils::success_output()),
-            }]);
-
-        let mut binstall_status = missing_link;
-        update_status_after_install(
-            &mut binstall_status,
-            &executor,
-            &CARGO_DYLINT_TOOL,
-            InstallOutcome::CargoBinstall,
-        );
-        assert!(binstall_status.cargo_dylint);
-        assert!(binstall_status.dylint_link);
-        executor.assert_finished();
-    }
-
-    #[test]
-    fn update_status_after_install_refreshes_link_for_cargo_install_outcome() {
-        let executor =
-            crate::test_utils::StubExecutor::new(vec![crate::test_utils::ExpectedCall {
-                cmd: "dylint-link",
-                args: vec!["--version"],
-                result: Ok(crate::test_utils::success_output()),
-            }]);
-        let mut status = DylintToolStatus {
-            cargo_dylint: false,
-            dylint_link: false,
-        };
-
-        update_status_after_install(
-            &mut status,
-            &executor,
-            &CARGO_DYLINT_TOOL,
-            InstallOutcome::CargoInstall,
-        );
-
-        assert!(status.cargo_dylint);
-        assert!(status.dylint_link);
-        executor.assert_finished();
-    }
-
-    #[test]
-    fn update_status_after_install_skips_link_probe_for_repository_release() {
-        let executor = crate::test_utils::StubExecutor::new(vec![]);
-        let mut status = DylintToolStatus {
-            cargo_dylint: false,
-            dylint_link: false,
-        };
-
-        update_status_after_install(
-            &mut status,
-            &executor,
-            &CARGO_DYLINT_TOOL,
-            InstallOutcome::RepositoryRelease,
-        );
-
-        assert!(status.cargo_dylint);
-        assert!(!status.dylint_link);
-        executor.assert_finished();
-    }
-
-    #[test]
-    fn should_install_tool_returns_true_for_cargo_dylint_when_not_installed() {
-        let status = DylintToolStatus {
-            cargo_dylint: false,
-            dylint_link: false,
-        };
-
-        assert!(should_install_tool(&status, &CARGO_DYLINT_TOOL));
-    }
-
-    #[test]
-    fn should_install_tool_returns_false_for_cargo_dylint_when_installed() {
-        let status = DylintToolStatus {
-            cargo_dylint: true,
-            dylint_link: false,
-        };
-
-        assert!(!should_install_tool(&status, &CARGO_DYLINT_TOOL));
-    }
-
-    #[test]
-    fn should_install_tool_returns_true_for_dylint_link_when_not_installed() {
-        let status = DylintToolStatus {
-            cargo_dylint: false,
-            dylint_link: false,
-        };
-
-        assert!(should_install_tool(&status, &DYLINT_LINK_TOOL));
-    }
-
-    #[test]
-    fn should_install_tool_returns_false_for_dylint_link_when_installed() {
-        let status = DylintToolStatus {
-            cargo_dylint: false,
-            dylint_link: true,
-        };
-
-        assert!(!should_install_tool(&status, &DYLINT_LINK_TOOL));
-    }
-
-    #[test]
-    fn should_refresh_companions_true_when_non_repo_release_and_link_missing() {
-        let status = DylintToolStatus {
-            cargo_dylint: false,
-            dylint_link: false,
-        };
-
-        assert!(should_refresh_companions(
-            InstallOutcome::CargoInstall,
-            &status
-        ));
-    }
-
-    #[test]
-    fn should_refresh_companions_false_for_repository_release() {
-        let status = DylintToolStatus {
-            cargo_dylint: false,
-            dylint_link: false,
-        };
-
-        assert!(!should_refresh_companions(
-            InstallOutcome::RepositoryRelease,
-            &status
-        ));
-    }
-
-    #[test]
-    fn should_refresh_companions_false_when_link_already_present() {
-        let status = DylintToolStatus {
-            cargo_dylint: true,
-            dylint_link: true,
-        };
-
-        assert!(!should_refresh_companions(
-            InstallOutcome::CargoBinstall,
-            &status
-        ));
-    }
-}
+#[path = "install_tests.rs"]
+mod tests;
